Pull JSDoc cache file processor out
