+++ src/resharper-angularjs/resharper-angularjs.csproj
@@ -163,8 +163,10 @@
   <ItemGroup>
     <Compile Include="Feature\Services\Caches\AngularJsCache.cs" />
     <Compile Include="Feature\Services\Caches\AngularJsCacheItems.cs" />
+    <Compile Include="Feature\Services\Caches\AngularJsCacheItemsBuilder.cs" />
     <Compile Include="Feature\Services\Caches\Directive.cs" />
     <Compile Include="Feature\Services\Caches\Filter.cs" />
+    <Compile Include="Feature\Services\Caches\JsDocFileProcessor.cs" />
     <Compile Include="Feature\Services\Caches\Parameter.cs" />
     <Compile Include="Feature\Services\CodeCompletion\AbbreviatedItemsProvider.cs" />
     <Compile Include="Feature\Services\CodeCompletion\AbbreviatedTextLookupItem.cs" />
